Move nilearn fetchers to functional (will split by type later).
