+++ nidata/_external/nilearn_old/nilearn/_utils/testing.py
@@ -19,7 +19,7 @@ import scipy.linalg
 from nibabel import Nifti1Image
 import nibabel
 
-from .. import datasets
+from .. import fetchers
 from .. import masking
 from . import logger
 
@@ -164,7 +164,7 @@ def mock_chunk_read_raise_error_(response, local_file, initial_size=0,
 
 
 class FetchFilesMock (object):
-    _mock_fetch_files = functools.partial(datasets._fetch_files, mock=True)
+    _mock_fetch_files = functools.partial(fetchers._fetch_files, mock=True)
 
     def __init__(self):
         """Create a mock that can fill a CSV file if needed
