+++ nidata/functional/description/craddock_2012.rst
