+++ nidata/functional/description/haxby2001.rst
