+++ nidata/functional/description/smith_2009.rst
